#56: moved tests to cli
